+++ plugins/sdk/src/lib/components/index.ts
@@ -1,2 +1,2 @@
-export * from './header';
+export * from './toolbar';
 export * from './sidebar';