+++ plugins/sdk/src/lib/services/navigation.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable, InjectionToken, Optional, Provider } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { AppHeaderEntry, AppSidebarEntry } from '../types';
+import { AppToolbarEntry, AppSidebarEntry } from '../types';
 
 export const APP_SIDEBAR_ENTRY_TOKEN = new InjectionToken<AppSidebarEntry>('Injection token for application sidebar entries.');
-export const APP_HEADER_ENTRY_TOKEN = new InjectionToken<AppHeaderEntry>('Injection token for application header entries.');
+export const APP_HEADER_ENTRY_TOKEN = new InjectionToken<AppToolbarEntry>('Injection token for application header entries.');
 
 /**
  * Generates injection providers to extend the application sidebar menu.
@@ -25,7 +25,7 @@ export function provideSidebarEntries(entries: AppSidebarEntry[]): Array<Provide
  * @param entries Header entries to register
  * @returns Module providers with the header entries
  */
-export function provideHeaderEntries(entries: AppHeaderEntry[]): Array<Provider> {
+export function provideHeaderEntries(entries: AppToolbarEntry[]): Array<Provider> {
   return entries.map((entry) => ({
     provide: APP_HEADER_ENTRY_TOKEN,
     multi: true,
@@ -36,13 +36,13 @@ export function provideHeaderEntries(entries: AppHeaderEntry[]): Array<Provider>
 @Injectable({ providedIn: 'root' })
 export class NavigationService {
   /** Application header entries */
-  headerEntries$: Observable<AppHeaderEntry[]>;
+  headerEntries$: Observable<AppToolbarEntry[]>;
 
   /** Application sidebar entries */
   sidebarEntries$: Observable<AppSidebarEntry[]>;
 
   constructor(
-    @Optional() @Inject(APP_HEADER_ENTRY_TOKEN) headerEntries: AppHeaderEntry[],
+    @Optional() @Inject(APP_HEADER_ENTRY_TOKEN) headerEntries: AppToolbarEntry[],
     @Optional() @Inject(APP_SIDEBAR_ENTRY_TOKEN) sidebarEntries: AppSidebarEntry[]
   ) {
     this.headerEntries$ = new BehaviorSubject(headerEntries).asObservable();