+++ src/app/app.module.ts
@@ -15,7 +15,7 @@ registerLocaleData(localeEs);
 
 import { provideSidebarEntries, provideHeaderEntries } from '@app/sdk';
 import { appSidebarEntries } from './app.sidebar';
-import { appHeaderEntries } from './app.header';
+import { appToolbarEntries } from './app.toolbar';
 
 @NgModule({
   imports: [
@@ -30,7 +30,7 @@ import { appHeaderEntries } from './app.header';
     })
   ],
   declarations: [AppComponent],
-  providers: [provideTranslations('app', 'assets'), provideHeaderEntries(appHeaderEntries), provideSidebarEntries(appSidebarEntries)],
+  providers: [provideTranslations('app', 'assets'), provideHeaderEntries(appToolbarEntries), provideSidebarEntries(appSidebarEntries)],
   bootstrap: [AppComponent]
 })
 export class AppModule {